Set up Bazel as project build system
In this commit, we set up the Bazel build system for Accelerando. At the same
time, to make it easier for editors with sane defaults to use this system, we
also add a Makefile with sensible targets. Thus, to build and test, it suffices
to just run `make` or `make all`, like one would expect.

While I have some reservations about using Bazel, especially when it comes to
documentation, I do find it quite nice to use, without a lot of boilerplate.